endpoint: Remove BuildMutex
Now that the build queue guarantees that no parallel builds of a particular
endpoint can occur. It is simple to remoev the build mutex by first draining
the build queue from the Buildable. Because the endpoint is already in
Disconnecting state, no further builds can be scheduled and draining is
sufficient.

Signed-off-by: Thomas Graf <thomas@cilium.io>